requirements.txt
openai>=1.30.0
httpx>=0.27.0
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
sentence-transformers>=2.7.0
rouge-score>=0.1.2
bert-score>=0.3.13
transformers>=4.44.0
torch>=2.3.0
tqdm>=4.66.0
tenacity>=8.2.0
scikit-learn>=1.4.0
